server/openclaw-integration/requirements.txt
# AOCN OpenClaw Integration — Python Dependencies
playwright==1.49.1
pyyaml>=6.0
aiohttp>=3.9
python-telegram-bot>=21.0
requests>=2.31
websockets>=12.0
